+++ app/src/input_manager.h
@@ -6,7 +6,6 @@
 #include <SDL2/SDL.h>
 
 #include "config.h"
-#include "common.h"
 #include "controller.h"
 #include "fps_counter.h"
 #include "scrcpy.h"
